+++ src/interface/src/app/app.component.spec.ts
@@ -1,7 +1,7 @@
 import { TopBarComponent } from './top-bar/top-bar.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { of } from 'rxjs';
-import { AuthService } from './auth.service';
+import { AuthService } from './services';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
