+++ src/interface/src/app/app.module.ts
@@ -11,17 +11,14 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { AccountDialogComponent } from './account-dialog/account-dialog.component';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { AuthGuard, AuthService } from './auth.service';
+import { AuthGuard, AuthService, MapService, PopupService, SessionService } from './services';
 import { LoginComponent } from './login/login.component';
-import { MapService } from './map.service';
 import { MapNameplateComponent } from './map/map-nameplate/map-nameplate.component';
 import { MapComponent } from './map/map.component';
 import { ProjectCardComponent } from './map/project-card/project-card.component';
 import { MaterialModule } from './material/material.module';
 import { NavigationComponent } from './navigation/navigation.component';
-import { PopupService } from './popup.service';
 import { RegionSelectionComponent } from './region-selection/region-selection.component';
-import { SessionService } from './session.service';
 import { SharedModule } from './shared/shared.module';
 import { SignupComponent } from './signup/signup.component';
 import { StringifyMapConfigPipe } from './stringify-map-config.pipe';
