+++ src/interface/src/app/login/login.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthService } from './../auth.service';
+import { AuthService } from '../services';
 
 @Component({
   selector: 'app-login',
