+++ src/interface/src/app/navigation/navigation.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { BehaviorSubject, of, Subject } from 'rxjs';
 
-import { AuthService } from './../auth.service';
+import { AuthService } from '../services';
 import { NavigationComponent } from './navigation.component';
 
 describe('NavigationComponent', () => {
