+++ src/interface/src/app/region-selection/region-selection.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { Region } from '../types';
 import { RegionSelectionComponent } from './region-selection.component';
-import { SessionService } from './../session.service';
+import { SessionService } from '../services';
 
 // selector names for getting DOM elements
 enum CssSelector {
