+++ src/interface/src/app/services/plan.service.ts
@@ -2,7 +2,7 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, catchError, map, Observable, of, take, tap } from 'rxjs';
 
-import { BasePlan, Plan } from './types';
+import { BasePlan, Plan } from '../types';
 
 export interface PlanState {
   all: {
