+++ src/interface/src/app/services/session.service.spec.ts
@@ -1,4 +1,4 @@
-import { Region } from './types/region.types';
+import { Region } from '../types/region.types';
 import { TestBed } from '@angular/core/testing';
 
 import { SessionService } from './session.service';
